Grant Totem access to the libvisual libraries it needs when using GStreamer 1.4.
This is needed on current Debian unstable.

Note that this will need to be moved to a gstreamer abstraction at some point,
as the binary accessing these libraries is actually gst-plugin-scanner.
I still intend to work on this, and am blocked by waiting an answer in the
"gstreamer abstraction" thread on the AppArmor mailing-list.